move check/020 entirely to gfonts spec and simplify code
(issue #1948)
