+++ Lib/fontbakery/specifications/googlefonts.py
@@ -195,7 +195,7 @@ def style(font):
   return None
 
 
-@condition(force=True)
+@condition
 def expected_os2_weight(style):
   """The weight name and the expected OS/2 usWeightClass value inferred from
   the style part of the font name
@@ -239,17 +239,6 @@ def expected_os2_weight(style):
 
   expected = GF_API_WEIGHTS[weight_name]
   return weight_name, expected
-
-
-@condition(force=True)
-def os2_weight_warn():
-  return {"style": "ExtraLight",
-          "value": 250,
-          "message": ("A value of 250 for OS/2 usWeightClass is acceptable for TTFs"
-                      " (but not for OTFs), because it won't auto-bold (and blur)"
-                      " in Windows GDI apps. However, since OTFs will, and because"
-                      " we'd like to have OTFs and TTFs be as consistent as possible,"
-                      " we'd prefer ExtraLight to be 275 in both cases.")}
 
 
 @check(
@@ -634,6 +623,31 @@ def com_google_fonts_check_019(ttFont):
   if not failed:
     yield PASS, ("No need to substitute copyright, registered and"
                  " trademark symbols in name table entries of this font.")
+
+
+@check(
+  id = 'com.google.fonts/check/020'
+)
+def com_google_fonts_check_020(font, ttFont, style):
+  """Checking OS/2 usWeightClass."""
+#  from fontbakery.specifications.shared_conditions import is_ttf
+
+  weight_name, expected_value = expected_os2_weight(style)
+  value = ttFont['OS/2'].usWeightClass
+
+  if value != expected_value:
+    if weight_name == 'ExtraLight' and value == 250:
+      yield WARN, ("A value of 250 for OS/2 usWeightClass is acceptable for TTFs"
+                   " (but not for OTFs), because it won't auto-bold (and blur)"
+                   " in Windows GDI apps. However, since OTFs will, and because"
+                   " we'd like to have OTFs and TTFs be as consistent as possible,"
+                   " we'd prefer ExtraLight to be 275 in both cases.")
+    else:
+      yield FAIL, ("OS/2 usWeightClass expected value for"
+                   " '{}' is {} but this font has"
+                   " {}.").format(weight_name, expected_value, value)
+  else:
+    yield PASS, "OS/2 usWeightClass value looks good!"
 
 
 @condition
